Disable chat nudges and add fractional chat pricing
- Set CHAT_CREDITS_PER_MESSAGE=0.2 in config/env example
- Disable nudge logic in ChatService (nudge=False for now)
- Deduct 0.2 credits per sent chat message via CreditService
- Convert credits, transaction amounts, job charges, and plan grants to NUMERIC(10,2)
- Return 402 Payment Required when the user has insufficient chat credits

+++ backend/app/dependencies.py
@@ -48,7 +48,7 @@ async def get_current_user(
 get_current_active_user = Depends(get_current_user)
 
 
-def require_credits(n: int) -> Callable:
+def require_credits(n: float) -> Callable:
     async def _check(current_user: User = Depends(get_current_user)) -> User:
         if current_user.credits < n:
             raise ForbiddenError(
